+++ src/utils/index.ts
@@ -1,4 +1,6 @@
 export * from "./Camera"
 export * from "./Figure"
+export * from "./Queue"
+export * from "./Stack"
 export * from "./State"
 export * from "./Utils"